Create giving season banner component
