+++ packages/lesswrong/components/ea-forum/giving-portal/EAGivingPortalPage.tsx
@@ -16,6 +16,7 @@ import {
   donationElectionTagId,
   eaGivingSeason23ElectionName,
   effectiveGivingTagId,
+  heroImageId,
   postsAboutElectionLink,
   setupFundraiserLink,
   timelineSpec,
@@ -281,8 +282,6 @@ const canonicalUrl = getSiteUrl() + "giving-portal";
 
 const pageDescription = "It's Giving season on the EA Forum. We're hosting a Donation Election, weekly themes, and more throughout November and December 2023.";
 
-const heroImage = "giving_portal_23_hero";
-
 const socialImageProps: CloudinaryPropsType = {
   dpr: "auto",
   ar: "16:9",
@@ -365,7 +364,7 @@ const EAGivingPortalPage = ({classes}: {classes: ClassesType}) => {
           canonicalUrl={canonicalUrl}
           ogUrl={canonicalUrl}
           description={pageDescription}
-          image={makeCloudinaryImageUrl(heroImage, socialImageProps)}
+          image={makeCloudinaryImageUrl(heroImageId, socialImageProps)}
         />
         <div className={classNames(classes.content, classes.mb20)} id="top">
           <div className={classNames(classes.h1, classes.center, classes.mt30)}>
@@ -488,7 +487,7 @@ const EAGivingPortalPage = ({classes}: {classes: ClassesType}) => {
             </div>
           </div>
         </div>
-        <CloudinaryImage2 publicId={heroImage} fullWidthHeader imgProps={{ h: "1200" }} />
+        <CloudinaryImage2 publicId={heroImageId} fullWidthHeader imgProps={{ h: "1200" }} />
         <div className={classes.sectionLight}>
           <div className={classes.content}>
             <div className={classNames(
