+++ packages/lesswrong/components/ea-forum/giving-portal/GivingSeasonHeader.tsx
@@ -6,37 +6,40 @@ import {
   styles as headerStyles,
   forumHeaderTitleSetting,
   forumShortTitleSetting,
+  EA_FORUM_HEADER_HEIGHT,
 } from "../../common/Header";
-import { cloudinaryCloudNameSetting } from "../../../lib/publicSettings";
-import { timelineSpec } from "../../../lib/eaGivingSeason";
-import { useCurrentTime } from "../../../lib/utils/timeUtil";
+import { makeCloudinaryImageUrl } from "../../common/CloudinaryImage2";
 import { lightbulbIcon } from "../../icons/lightbulbIcon";
+import { heroImageId } from "../../../lib/eaGivingSeason";
+import { isEAForum } from "../../../lib/instanceSettings";
 import Typography from "@material-ui/core/Typography";
 import Toolbar from "@material-ui/core/Toolbar";
 import Headroom from "../../../lib/react-headroom";
 import classNames from "classnames";
-import moment from "moment";
-import { isEAForum } from "../../../lib/instanceSettings";
 
-const cloudinaryCloudName = cloudinaryCloudNameSetting.get();
-const gsImg = `https://res.cloudinary.com/${cloudinaryCloudName}/image/upload/c_crop,g_custom/h_300,q_100,f_auto/giving_portal_23_hero`;
+const gsImg = makeCloudinaryImageUrl(heroImageId, {
+  h: "300",
+  q: "100",
+  f: "auto",
+});
 
-const EA_FORUM_GIVING_SEASON_HEADER_HEIGHT = 213;
+export const EA_FORUM_GIVING_SEASON_HEADER_HEIGHT = 213;
 
 const styles = (theme: ThemeType) => ({
   ...headerStyles(theme),
   rootGivingSeason: {
-    height: EA_FORUM_GIVING_SEASON_HEADER_HEIGHT,
     "& .headroom": {
       zIndex: theme.zIndexes.searchResults,
     },
   },
   appBarGivingSeason: {
     color: theme.palette.givingPortal.homepageHeader.light4,
-    background: `center no-repeat url(${gsImg}), ${theme.palette.givingPortal.homepageHeader.dark}`,
+    background: [
+      `center no-repeat url(${gsImg})`,
+      theme.palette.givingPortal.homepageHeader.dark,
+    ],
     position: "static",
     width: "100%",
-    height: EA_FORUM_GIVING_SEASON_HEADER_HEIGHT,
     display: "flex",
     zIndex: 1100,
     boxSizing: "border-box",
@@ -45,7 +48,10 @@ const styles = (theme: ThemeType) => ({
     padding: "1px 20px",
     overflow: "hidden",
     "@media (max-width: 1200px)": {
-      background: `right no-repeat url(${gsImg}), ${theme.palette.givingPortal.homepageHeader.dark}`,
+      background: [
+        `right no-repeat url(${gsImg})`,
+        theme.palette.givingPortal.homepageHeader.dark,
+      ],
     },
     [theme.breakpoints.down("sm")]: {
       background: theme.palette.givingPortal.homepageHeader.main,
@@ -124,6 +130,7 @@ const styles = (theme: ThemeType) => ({
     color: theme.palette.givingPortal.homepageHeader.light4,
   },
   givingSeasonGradient: {
+    display: "none",// TMP TODO
     position: "absolute",
     top: 0,
     left: 0,
@@ -136,120 +143,6 @@ const styles = (theme: ThemeType) => ({
     "@media (max-width: 1200px)": {
       background: `linear-gradient(76deg, ${theme.palette.givingPortal.homepageHeader.dark} 10%, ${theme.palette.givingPortal.homepageHeader.main} 40%, ${theme.palette.background.transparent} 70%, ${theme.palette.givingPortal.homepageHeader.main})`,
     },
-    [theme.breakpoints.down("sm")]: {
-      display: "none",
-    },
-  },
-  givingSeasonContent: {
-    display: "flex",
-    alignItems: "center",
-    marginTop: 16,
-    zIndex: theme.zIndexes.spotlightItem,
-    [theme.breakpoints.down("sm")]: {
-      flexDirection: "column",
-      alignItems: "flex-start",
-      gap: "16px",
-      height: "100%",
-      marginTop: 0,
-    },
-  },
-  givingSeasonOverview: {
-    padding: "0 20px 0 48px",
-    flexGrow: 1,
-    [theme.breakpoints.down("sm")]: {
-      padding: 0,
-    },
-  },
-  givingSeasonHeading: {
-    color: theme.palette.givingPortal.homepageHeader.light4,
-    fontSize: 40,
-    lineHeight: "48px",
-    marginTop: 0,
-    marginBottom: 8,
-    [theme.breakpoints.down("sm")]: {
-      marginTop: 12,
-      marginBottom: 16,
-      fontSize: 30,
-      lineHeight: "30px",
-    },
-  },
-  givingSeasonDescription: {
-    color: theme.palette.givingPortal.homepageHeader.light4,
-    paddingLeft: 3,
-    maxWidth: 500,
-    [theme.breakpoints.down("sm")]: {
-      fontSize: 13,
-    },
-  },
-  givingSeasonLink: {
-    textDecoration: "underline",
-  },
-  givingSeasonTimeline: {
-    maxWidth: 600,
-    display: "grid",
-    gridTemplateColumns: "repeat(3, max-content)",
-    alignItems: "center",
-    textAlign: "center",
-    fontFamily: theme.palette.fonts.sansSerifStack,
-    color: theme.palette.givingPortal.homepageHeader.light4,
-    fontSize: 13,
-    fontWeight: 500,
-    marginRight: -20,
-    "@media (max-width: 1400px)": {
-      maxWidth: 480,
-      gridTemplateColumns: "repeat(3, auto)",
-    },
-    [theme.breakpoints.down("sm")]: {
-      alignSelf: "flex-end",
-      marginBottom: 16,
-    },
-    ["@media (max-width: 600px)"]: {
-      marginBottom: 12,
-    },
-    ["@media (max-width: 500px)"]: {
-      alignSelf: "flex-start",
-      marginBottom: 12,
-    },
-    ["@media (max-width: 280px)"]: {
-      display: "none",
-    },
-  },
-  gsTimelineLabel: {
-    backgroundColor: theme.palette.givingPortal.homepageHeader.secondaryOpaque,
-    padding: "6px 12px",
-    borderLeft: `1px solid ${theme.palette.givingPortal.homepageHeader.main}`,
-    "&:first-of-type": {
-      borderLeft: "none"
-    },
-    "&:hover": {
-      backgroundColor: theme.palette.givingPortal.homepageHeader.secondaryOpaqueDark,
-      opacity: 1
-    },
-    [theme.breakpoints.down("sm")]: {
-      whiteSpace: "nowrap",
-      padding: "4px 8px",
-    },
-  },
-  gsTimelineLabelActive: {
-    backgroundColor: theme.palette.givingPortal.homepageHeader.light2,
-    color: theme.palette.givingPortal.homepageHeader.main,
-    fontSize: 16,
-    fontWeight: 600,
-    padding: "10px 20px",
-    borderRadius: theme.borderRadius.default,
-    borderLeft: "none",
-    "&:hover": {
-      backgroundColor: theme.palette.givingPortal.homepageHeader.light1,
-      opacity: 1
-    },
-    [theme.breakpoints.down("sm")]: {
-      fontSize: 13,
-      padding: "8px 10px",
-      borderRadius: theme.borderRadius.default/2,
-    },
-  },
-  gsTimelineDates: {
-    paddingTop: 8,
     [theme.breakpoints.down("sm")]: {
       display: "none",
     },
@@ -275,20 +168,18 @@ const GivingSeasonHeader = ({
   HeaderNotificationsMenu: FC,
   classes: ClassesType,
 }) => {
-  const now = useCurrentTime();
   return (
     <AnalyticsContext pageSectionContext="header" siteEvent="givingSeason2023">
-      <div className={classes.rootGivingSeason}>
+      <div className={classNames(classes.root, classes.rootGivingSeason)}>
         <Headroom
           disableInlineStyles
           downTolerance={10} upTolerance={10}
-          height={EA_FORUM_GIVING_SEASON_HEADER_HEIGHT}
+          height={EA_FORUM_HEADER_HEIGHT}
           className={classNames(classes.headroom, {
             [classes.headroomPinnedOpen]: searchOpen,
           })}
           onUnfix={() => setUnFixed(true)}
           onUnpin={() => setUnFixed(false)}
-          disable={false}
         >
           <header className={classes.appBarGivingSeason}>
             <div className={classes.givingSeasonGradient}></div>
@@ -323,36 +214,6 @@ const GivingSeasonHeader = ({
               </Typography>
               <RightHeaderItems />
             </Toolbar>
-            <div className={classes.givingSeasonContent}>
-              <div className={classes.givingSeasonOverview}>
-                <Typography variant="display1" className={classes.givingSeasonHeading}>
-                  Giving season 2023
-                </Typography>
-                <Typography variant="body2" className={classes.givingSeasonDescription} component="div">
-                  Donate to the Election Fund and discuss where the donations should go. <Link to="/giving-portal" className={classes.givingSeasonLink}>Learn more in the Giving portal.</Link>
-                </Typography>
-              </div>
-              <div className={classes.givingSeasonTimeline}>
-                {timelineSpec.spans.map((span) => {
-                  // ignore the voting time period
-                  if (span.hatched) return null
-                  const isActive = moment.utc(span.start).isBefore(now) && moment.utc(span.end).isAfter(now)
-                  return <Link
-                    key={`${span.description}-label`}
-                    to={span.href ?? "#"}
-                    className={classNames(classes.gsTimelineLabel, {[classes.gsTimelineLabelActive]: isActive})}
-                  >
-                    {span.description}
-                  </Link>
-                })}
-                {timelineSpec.spans.map(span => {
-                  if (span.hatched) return null
-                  return <div key={`${span.description}-dates`} className={classes.gsTimelineDates}>
-                    {moment.utc(span.start).format("MMM D")}-{moment.utc(span.end).format("D")}
-                  </div>
-                })}
-              </div>
-            </div>
           </header>
           <HeaderNavigationDrawer />
         </Headroom>
