+++ packages/lesswrong/lib/eaGivingSeason.ts
@@ -12,6 +12,8 @@ export const votingOpensDate = new Date("2023/12/01");
 export const donationElectionTagId = "EsNWGoFbs4MrqQ4G7";
 export const effectiveGivingTagId = "L6NqHZkLc4xZ7YtDr";
 
+export const heroImageId = "giving_portal_23_hero";
+
 const votingAccountCreationCutoff = new Date("2023/10/23");
 
 const userCanVoteInDonationElection = (
